# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  version: 1.0.0
  description: Canonical contract for Buyer, Seller and Admin clients.
  license:
    name: Proprietary
    url: https://marketplace.local/legal/license
servers:
  - url: /api/v1
security: []
...
paths:
...
  /products:
    get:
      tags: [Products]
      operationId: searchProducts
      summary: Search active products
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { name: query, in: query, schema: { type: string, maxLength: 200 } }
        - { name: category, in: query, schema: { type: string } }
        - { name: brand, in: query, schema: { type: string } }
        - { name: type, in: query, schema: { $ref: '#/components/schemas/ProductType' } }
        - { name: minPrice, in: query, schema: { type: integer, format: int64, minimum: 0 } }
        - { name: maxPrice, in: query, schema: { type: integer, format: int64, minimum: 0 } }
        - { name: sortBy, in: query, schema: { type: string, enum: [RELEVANCE, NEWEST, PRICE_ASC, PRICE_DESC, BEST_SELLER, RATING] } }
      responses:
        '200':
          description: Cursor-paginated products
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    required: [data]
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/ProductSummary' } }
        '400': { $ref: '#/components/responses/BadRequest' }
  /products/{slug}:
    get:
      tags: [Products]
      operationId: getProductBySlug
      summary: Get a public product detail
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Public product detail without inventory secrets
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/ProductDetail' } }
        '404': { $ref: '#/components/responses/NotFound' }
  /cart:
    get:
      tags: [Cart]
      security: [{ bearerAuth: [] }]
      operationId: getCart
      summary: Get the current buyer cart
      responses:
        '200':
          description: Current cart priced for display only
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /cart/items:
    post:
      tags: [Cart]
      security: [{ bearerAuth: [] }]
      operationId: addCartItem
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddCartItemRequest' }
      responses:
        '201': { description: Item added }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /cart/items/{id}:
    patch:
      tags: [Cart]
      security: [{ bearerAuth: [] }]
      operationId: updateCartItem
      summary: Update a cart item
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                quantity: { type: integer, minimum: 1, maximum: 99 }
                selected: { type: boolean }
              minProperties: 1
      responses:
        '200': { description: Updated }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Cart]
      security: [{ bearerAuth: [] }]
      operationId: removeCartItem
      summary: Remove a cart item
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '204': { description: Removed }
        '404': { $ref: '#/components/responses/NotFound' }
  /checkout/preview:
    post:
      tags: [Checkout]
      security: [{ bearerAuth: [] }]
      operationId: previewCheckout
      summary: Price selected cart items
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckoutPreviewRequest' }
      responses:
        '200':
          description: Server-priced snapshot with short TTL
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/CheckoutPreview' } }
        '409': { $ref: '#/components/responses/Conflict' }
  /checkout/order-groups:
    post:
      tags: [Checkout]
      security: [{ bearerAuth: [] }]
      operationId: createOrderGroup
      summary: Create orders and a payment intent atomically
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderGroupRequest' }
      responses:
        '201':
          description: Atomic order group and payment intent creation
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/CreateOrderGroupResult' } }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /payments/{id}:
    get:
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      operationId: getPayment
      summary: Get a payment intent
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Owner/admin-safe payment detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/Payment' } }
        '404': { $ref: '#/components/responses/NotFound' }
  /payments/{id}/status:
    get:
      tags: [Payments]
      security: [{ bearerAuth: [] }]
      operationId: getPaymentStatus
      summary: Poll payment status
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: Lightweight polling response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [status, isPaid]
                        properties:
                          status: { $ref: '#/components/schemas/PaymentStatus' }
                          isPaid: { type: boolean }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /webhooks/sepay:
    post:
      tags: [Payments]
      operationId: receiveSepayWebhook
      summary: Receive a SePay transaction event
      description: Provider-authenticated endpoint. Raw body is required for signature schemes.
      security: [{ sepayWebhookAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200': { description: 'Durably accepted, including known duplicates' }
        '401': { description: Invalid provider authentication }
  /orders:
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      operationId: listOrders
      summary: List orders visible to the current actor
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/OrderStatus' } }
      responses:
        '200':
          description: Orders scoped to current actor
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: '#/components/schemas/Order' } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /orders/{id}:
    get:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      operationId: getOrder
      summary: Get an order detail
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { description: Order detail with server-computed allowedActions }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{id}/cancel:
    post:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      operationId: cancelOrder
      summary: Cancel an eligible order
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties: { reason: { type: string, minLength: 3, maxLength: 500 } }
      responses:
        '200': { description: Cancelled or prior idempotent result }
        '409': { $ref: '#/components/responses/Conflict' }
  /orders/{id}/confirm-received:
    post:
      tags: [Orders]
      security: [{ bearerAuth: [] }]
      operationId: confirmOrderReceived
      summary: Confirm order receipt
      parameters:
        - { $ref: '#/components/parameters/Id' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200': { description: Completed or prior idempotent result }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sepayWebhookAuth:
      type: apiKey
      in: header
      name: Authorization
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$' }
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 500 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
  responses:
...
    BadRequest: { description: Validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unauthorized: { description: Authentication failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
    NotFound: { description: Not found or hidden by ownership policy, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: 'State, inventory, version or idempotency conflict', content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Unprocessable: { description: Business validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data, meta]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta:
          type: object
          required: [requestId]
          properties:
            requestId: { type: string }
            nextCursor: { type: string, nullable: true }
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, example: INVENTORY_OUT_OF_STOCK }
            message: { type: string }
            details: { type: object, additionalProperties: true }
            requestId: { type: string }
...
    ProductType:
      type: string
      enum: [PHYSICAL, DIGITAL_GAME_ACCOUNT, DIGITAL_APP_ACCOUNT]
    ProductSummary:
      type: object
      required: [id, slug, name, type, minPrice, currency, availability]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        type: { $ref: '#/components/schemas/ProductType' }
        minPrice: { type: integer, format: int64, minimum: 0 }
        currency: { type: string, enum: [VND] }
        availability: { type: string, enum: [IN_STOCK, LOW_STOCK, OUT_OF_STOCK] }
    ProductDetail:
      allOf:
        - $ref: '#/components/schemas/ProductSummary'
        - type: object
          required: [description, shop, purchaseOptions]
          properties:
            description: { type: string }
            shop: { type: object, additionalProperties: true }
            attributes: { type: object, additionalProperties: true }
            purchaseOptions: { type: array, items: { type: object, additionalProperties: true } }
    AddCartItemRequest:
      oneOf:
        - $ref: '#/components/schemas/AddPhysicalCartItem'
        - $ref: '#/components/schemas/AddGameCartItem'
        - $ref: '#/components/schemas/AddAppCartItem'
      discriminator: { propertyName: productType }
    AddPhysicalCartItem:
      type: object
      additionalProperties: false
      required: [productId, productType, productVariantId, quantity]
      properties:
        productId: { type: string, format: uuid }
        productType: { type: string, enum: [PHYSICAL] }
        productVariantId: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 99 }
    AddGameCartItem:
      type: object
      additionalProperties: false
      required: [productId, productType, quantity]
      properties:
        productId: { type: string, format: uuid }
        productType: { type: string, enum: [DIGITAL_GAME_ACCOUNT] }
        quantity: { type: integer, minimum: 1, maximum: 10 }
    AddAppCartItem:
      type: object
      additionalProperties: false
      required: [productId, productType, appPlanId, quantity]
      properties:
        productId: { type: string, format: uuid }
        productType: { type: string, enum: [DIGITAL_APP_ACCOUNT] }
        appPlanId: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 10 }
        buyerProvidedValues: { type: object, additionalProperties: { type: string } }
    CartItem:
      type: object
      required: [id, productId, productType, quantity, selected]
      properties:
        id: { type: string, format: uuid }
        productId: { type: string, format: uuid }
        productType: { $ref: '#/components/schemas/ProductType' }
        productVariantId: { type: string, format: uuid, nullable: true }
        appPlanId: { type: string, format: uuid, nullable: true }
        quantity: { type: integer }
        selected: { type: boolean }
    CheckoutPreviewRequest:
      type: object
      additionalProperties: false
      required: [selectedCartItemIds]
      properties:
        selectedCartItemIds:
          type: array
          minItems: 1
          uniqueItems: true
          items: { type: string, format: uuid }
        addressId: { type: string, format: uuid, nullable: true }
        shippingMethodId: { type: string, nullable: true }
        voucherCodes: { type: array, maxItems: 10, items: { type: string } }
    CheckoutPreview:
      type: object
      required: [checkoutSessionId, expiresAt, subtotal, discount, shipping, tax, total, currency, shopBreakdown]
      properties:
        checkoutSessionId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        subtotal: { type: integer, format: int64 }
        discount: { type: integer, format: int64 }
        shipping: { type: integer, format: int64 }
        tax: { type: integer, format: int64 }
        total: { type: integer, format: int64 }
        currency: { type: string, enum: [VND] }
        shopBreakdown: { type: array, items: { type: object, additionalProperties: true } }
    CreateOrderGroupRequest:
      type: object
      additionalProperties: false
      required: [checkoutSessionId, paymentMethod]
      properties:
        checkoutSessionId: { type: string, format: uuid }
        paymentMethod: { type: string, enum: [SEPAY_QR, COD] }
        notes: { type: string, maxLength: 1000 }
    CreateOrderGroupResult:
      type: object
      required: [orderGroupId, orderIds, paymentId, totalAmount]
      properties:
        orderGroupId: { type: string, format: uuid }
        orderIds: { type: array, items: { type: string, format: uuid } }
        paymentId: { type: string, format: uuid }
        totalAmount: { type: integer, format: int64 }
    PaymentStatus:
      type: string
      enum: [CREATED, PENDING, PAID, EXPIRED, REVIEW_REQUIRED, REJECTED, PARTIALLY_REFUNDED, REFUNDED]
    Payment:
      type: object
      required: [id, orderGroupId, amount, currency, status, expiresAt]
      properties:
        id: { type: string, format: uuid }
        orderGroupId: { type: string, format: uuid }
        amount: { type: integer, format: int64 }
        currency: { type: string, enum: [VND] }
        status: { $ref: '#/components/schemas/PaymentStatus' }
        expiresAt: { type: string, format: date-time }
        paidAt: { type: string, format: date-time, nullable: true }
        sepayInfo: { type: object, additionalProperties: true }
    OrderStatus:
      type: string
      enum: [PENDING_PAYMENT, PAID, PROCESSING, DELIVERED, COMPLETED, EXPIRED, CANCELLED, DISPUTED, REFUNDED]
    Order:
      type: object
      required: [id, code, shopId, status, totalAmount, currency, items, allowedActions]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        shopId: { type: string, format: uuid }
        status: { $ref: '#/components/schemas/OrderStatus' }
        totalAmount: { type: integer, format: int64 }
        currency: { type: string, enum: [VND] }
        items: { type: array, items: { type: object, additionalProperties: true } }
        allowedActions: { type: array, items: { type: string } }
...